requirements.txt
addict==2.2.1
albumentations==0.4.3
einops==0.3.0
matplotlib==3.5.1
nystrom-attention==0.0.9
opencv-python==4.2.0.34
opencv-python-headless==4.2.0.34
pandas==1.2.3
Pillow==8.4.0
pytorch-lightning==1.2.3
pytorch-toolbelt==0.4.0
torchmetrics==0.4.1
